moved HistoryComponent HOC to UIElements.js file
