+++ poem/Poem/frontend/react/Probes.js
@@ -1,7 +1,15 @@
 import React, { Component } from 'react';
 import { Backend } from './DataManager';
 import { Link } from 'react-router-dom';
-import { LoadingAnim, BaseArgoView, NotifyOk, Checkbox, FancyErrorMessage, AutocompleteField } from './UIElements';
+import { 
+  LoadingAnim, 
+  BaseArgoView, 
+  NotifyOk, 
+  Checkbox, 
+  FancyErrorMessage, 
+  AutocompleteField, 
+  HistoryComponent
+} from './UIElements';
 import ReactTable from 'react-table';
 import {
   FormGroup, 
@@ -18,7 +26,7 @@ import { NotificationManager } from 'react-notifications';
 import * as Yup from 'yup';
 
 
-export const ProbeHistory = HistoryComponent('probe')
+export const ProbeHistory = HistoryComponent('probe');
 
 
 export const DiffElement = ({title, item1, item2}) => {
@@ -814,155 +822,6 @@ export class ProbeChange extends Component {
     };
   };
 };
-
-
-export function HistoryComponent(obj) {
-  return class extends Component {
-    constructor(props) {
-      super(props);
-
-      this.name = props.match.params.name;
-      this.history = props.history;
-
-      this.state = {
-        loading: false,
-        list_versions: null,
-        compare1: '',
-        compare2: ''
-      };
-
-      this.backend = new Backend();
-    }
-
-    componentDidMount() {
-      this.setState({loading: true});
-      let url = undefined;
-
-      if (obj === 'metric')
-        url = '/api/v2/internal/tenantversion/'
-      
-      else
-        url = '/api/v2/internal/version/'
-
-      this.backend.fetchData(`${url}/${obj}/${this.name}`)
-        .then((json) => {
-          if (json.length > 1) {
-            this.setState({
-              list_versions: json,
-              loading: false,
-              compare1: json[0].version,
-              compare2: json[1].version
-            });
-          } else {
-            this.setState({
-              list_versions: json,
-              loading: false
-            });
-          };
-        }
-      )
-    };
-
-      render() {
-        const { loading, list_versions } = this.state; 
-    
-        if (loading)
-          return (<LoadingAnim />);
-        
-        else if (!loading && list_versions) {
-          return (
-            <BaseArgoView
-              resourcename='Change history'
-              infoview={true}>
-                <table className='table table-sm'>
-                  <thead className='table-active'>
-                    <tr>
-                      { list_versions.length === 1 ?
-                        <th scope='col'>Compare</th>
-                      :
-                        <th scope='col'>
-                          <Button
-                            color='info'
-                            onClick={() => 
-                              this.history.push(
-                                '/ui/'+obj+'s/' + this.name + '/history/compare/' + this.state.compare1 + '/' + this.state.compare2,
-                              )
-                            }
-                          >
-                            Compare
-                          </Button>
-                        </th>
-                      }
-                      <th scope='col'>Version</th>
-                      <th scope='col'>Date/time</th>
-                      <th scope='col'>User</th>
-                      <th scope='col'>Comment</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {
-                      list_versions.map((e, i) =>
-                        <tr key={i}>
-                          {
-                            list_versions.length === 1 ?
-                              <td>-</td>
-                            :
-                              i === 0 ?
-                              <td>
-                                <input
-                                  type='radio'
-                                  name='radio-1'
-                                  value={e.version}
-                                  defaultChecked={true}
-                                  onChange={e => this.setState({compare1: e.target.value})}
-                                />
-                              </td>
-                              :
-                              <td>
-                                <input
-                                  type='radio'
-                                  name='radio-1'
-                                  value={e.version}
-                                  onChange={e => this.setState({compare1: e.target.value})}
-                                /> 
-                                {' '}
-                                <input
-                                  type='radio'
-                                  name='radio-2'
-                                  value={e.version}
-                                  defaultChecked={i===1}
-                                  onChange={e => this.setState({compare2: e.target.value})}
-                                />
-                              </td>
-                          }
-                          {
-                            <td>
-                              {e.version ? <Link to={'/ui/' + obj +'s/' + this.name + '/history/' + e.version}>{e.version}</Link> : ''}
-                            </td>
-                          }
-                          <td>
-                            {e.date_created ? e.date_created : ''}
-                          </td>
-                          <td>
-                            {e.user ? e.user : ''}
-                          </td>
-                          <td className='col-md-6'>
-                            {e.comment ? e.comment : ''}
-                          </td>
-                        </tr>
-                      )
-                    }
-                  </tbody>
-                </table>
-              </BaseArgoView>
-          );
-        }
-        else
-          return null;
-      };
-    };
-};
-
 
 
 export class ProbeVersionCompare extends Component{
