+++ poem/Poem/frontend/react/UIElements.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Component } from 'react';
 import Cookies from 'universal-cookie';
 import {
   Alert,
@@ -47,6 +47,7 @@ import {
 import { NotificationManager } from 'react-notifications';
 import { Field } from 'formik';
 import Autocomplete from 'react-autocomplete';
+import { Backend } from './DataManager';
 
 
 var list_pages = ['administration','services', 'reports', 'probes',
@@ -539,3 +540,151 @@ export const DropdownFilterComponent = ({value, onChange, data}) => (
     }
   </select>
 )
+
+
+export function HistoryComponent(obj) {
+  return class extends Component {
+    constructor(props) {
+      super(props);
+
+      this.name = props.match.params.name;
+      this.history = props.history;
+
+      this.state = {
+        loading: false,
+        list_versions: null,
+        compare1: '',
+        compare2: ''
+      };
+
+      this.backend = new Backend();
+    }
+
+    componentDidMount() {
+      this.setState({loading: true});
+      let url = undefined;
+
+      if (obj === 'metric')
+        url = '/api/v2/internal/tenantversion/'
+      
+      else
+        url = '/api/v2/internal/version/'
+
+      this.backend.fetchData(`${url}/${obj}/${this.name}`)
+        .then((json) => {
+          if (json.length > 1) {
+            this.setState({
+              list_versions: json,
+              loading: false,
+              compare1: json[0].version,
+              compare2: json[1].version
+            });
+          } else {
+            this.setState({
+              list_versions: json,
+              loading: false
+            });
+          };
+        }
+      )
+    };
+
+      render() {
+        const { loading, list_versions } = this.state; 
+    
+        if (loading)
+          return (<LoadingAnim />);
+        
+        else if (!loading && list_versions) {
+          return (
+            <BaseArgoView
+              resourcename='Version history'
+              infoview={true}>
+                <table className='table table-sm'>
+                  <thead className='table-active'>
+                    <tr>
+                      { list_versions.length === 1 ?
+                        <th scope='col'>Compare</th>
+                      :
+                        <th scope='col'>
+                          <Button
+                            color='info'
+                            onClick={() => 
+                              this.history.push(
+                                '/ui/'+obj+'s/' + this.name + '/history/compare/' + this.state.compare1 + '/' + this.state.compare2,
+                              )
+                            }
+                          >
+                            Compare
+                          </Button>
+                        </th>
+                      }
+                      <th scope='col'>Version</th>
+                      <th scope='col'>Date/time</th>
+                      <th scope='col'>User</th>
+                      <th scope='col'>Comment</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    {
+                      list_versions.map((e, i) =>
+                        <tr key={i}>
+                          {
+                            list_versions.length === 1 ?
+                              <td>-</td>
+                            :
+                              i === 0 ?
+                              <td>
+                                <input
+                                  type='radio'
+                                  name='radio-1'
+                                  value={e.version}
+                                  defaultChecked={true}
+                                  onChange={e => this.setState({compare1: e.target.value})}
+                                />
+                              </td>
+                              :
+                              <td>
+                                <input
+                                  type='radio'
+                                  name='radio-1'
+                                  value={e.version}
+                                  onChange={e => this.setState({compare1: e.target.value})}
+                                /> 
+                                {' '}
+                                <input
+                                  type='radio'
+                                  name='radio-2'
+                                  value={e.version}
+                                  defaultChecked={i===1}
+                                  onChange={e => this.setState({compare2: e.target.value})}
+                                />
+                              </td>
+                          }
+                          {
+                            <td>
+                              {e.version ? <Link to={'/ui/' + obj +'s/' + this.name + '/history/' + e.version}>{e.version}</Link> : ''}
+                            </td>
+                          }
+                          <td>
+                            {e.date_created ? e.date_created : ''}
+                          </td>
+                          <td>
+                            {e.user ? e.user : ''}
+                          </td>
+                          <td className='col-md-6'>
+                            {e.comment ? e.comment : ''}
+                          </td>
+                        </tr>
+                      )
+                    }
+                  </tbody>
+                </table>
+              </BaseArgoView>
+          );
+        }
+        else
+          return null;
+      };
+    };
+};
